Moved checksummer to  separate file
